refactor: simplify internal service imports and extract classes
